bugfix: use float instead of int for modMenu/listMenus API calls, as this was corrupting the coordinates of sprites on uimenus

+++ src/IndexSprite.cpp
@@ -12,9 +12,13 @@ sf::Sprite IndexSprite::drawable()
     s.setTexture(tex.texture);
     s.setScale(sf::Vector2f(scale, scale));
     s.setPosition(sf::Vector2f(x, y));
+    //std::cout << "Drawable position: " << x << ", " << y << " -> " << s.getPosition().x << ", " << s.getPosition().y << std::endl;
     s.setTextureRect(tex.spriteCoordinates(sprite_index));
     return s;
 }
 
-IndexSprite::IndexSprite(int _ti, int _si, int _x, int _y, float _s):
-    texture_index(_ti), sprite_index(_si), x(_x), y(_y), scale(_s) {}
+IndexSprite::IndexSprite(int _ti, int _si, float _x, float _y, float _s):
+    texture_index(_ti), sprite_index(_si), x(_x), y(_y), scale(_s) {
+		//std::cout << "IndexSprite constructed with x, y " << _x << ", " << _y << std::endl;
+		//std::cout << "  * Stored x, y " << x << ", " << y << std::endl;
+}
